+++ plotnine/tests/test_facets.py
@@ -5,7 +5,7 @@ import pandas as pd
 import pytest
 
 from plotnine import ggplot, aes, geom_point, facet_grid, facet_wrap
-from plotnine import geom_abline, theme
+from plotnine import geom_abline
 
 n = 10
 df = pd.DataFrame({'x': range(n),
@@ -86,29 +86,22 @@ def test_facet_grid_margins():
 
 
 def test_facet_grid_scales_free_y():
-    p = (g
-         + facet_grid(['.', 'var1>2'], scales='free_y')
-         + theme(panel_spacing_x=0.3))
+    p = g + facet_grid('var1>2 ~ x%2', scales='free_y')
     assert p == 'facet_grid_scales_free_y'
 
 
-def test_facet_grid_scales_free_y_formula_dot_notation():
-    p = (g+facet_grid('. ~ var1>2', scales='free_y')
-         + theme(panel_spacing_x=0.3))
-    assert p == 'facet_grid_scales_free_y'
+def test_facet_grid_formula_with_dot():
+    p = g + facet_grid('. ~ var1>2')
+    assert p == 'facet_grid_formula_with_dot'
 
 
-def test_facet_grid_scales_free_y_formula_no_dot():
-    p = (g
-         + facet_grid('~var1>2', scales='free_y')
-         + theme(panel_spacing_x=0.3))
-    assert p == 'facet_grid_scales_free_y'
+def test_facet_grid_formula_without_dot():
+    p = g + facet_grid('~var1>2')
+    assert p == 'facet_grid_formula_with_dot'
 
 
 def test_facet_grid_scales_free_x():
-    p = (g
-         + facet_grid(['var1>2', '.'], scales='free_x')
-         + theme(panel_spacing_y=0.3))
+    p = g + facet_grid('var1>2 ~ x%2', scales='free_x')
     assert p == 'facet_grid_scales_free_x'